ASoC: SOF: Rename compress.c -> probe_compress.c
This will help separate compress probe helpers functions from
compress audio implementation that will follow in the next patches.
