+++ sound/soc/sof/intel/hda-dai.c
@@ -434,7 +434,7 @@ static const struct snd_soc_dai_ops hda_link_dai_ops = {
 };
 
 #if IS_ENABLED(CONFIG_SND_SOC_SOF_HDA_PROBES)
-#include "../compress.h"
+#include "../probe_compress.h"
 
 static struct snd_soc_cdai_ops sof_probe_compr_ops = {
 	.startup	= sof_probe_compr_open,
